:recycle: refactor(api): move findUserLikedItemsPoints ItemLocationRepository -> ItemLikeRepositoryImpl
