+++ backend/streetdrop-api/src/main/java/com/depromeet/domains/item/repository/ItemLikeRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ItemLikeRepository extends JpaRepository<ItemLike, Long> {
+public interface ItemLikeRepository extends JpaRepository<ItemLike, Long>, QueryDelItemLikeRepository {
     Optional<ItemLike> findByItemIdAndUser(Long itemId, User user);
 
     boolean existsByUserIdAndItemId(Long userId, Long itemId);
