+++ backend/streetdrop-api/src/main/java/com/depromeet/domains/item/repository/QueryDelItemLikeRepository.java
@@ -1,0 +1,9 @@
+package com.depromeet.domains.item.repository;
+
+import com.depromeet.domains.user.dao.UserItemPointDao;
+
+import java.util.List;
+
+public interface QueryDelItemLikeRepository {
+    List<UserItemPointDao> findUserLikedItemsPoints(Long userId);
+}
